requirements.txt
Flask>=2.2.5
Flask-SQLAlchemy>=3.0.5
Werkzeug>=2.2.3
psycopg2-binary>=2.9.9
numpy>=1.24.4
pandas>=1.5.3
matplotlib>=3.7.1
seaborn>=0.12.2
